Extract statements_explain.go
